requirements.txt
aiofiles==0.8.0
aiohttp==3.9.5
aiosignal==1.3.1
aiosqlite==0.17.0
astroid==3.2.4
asyncpraw==7.7.1
asyncprawcore==2.4.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
dill==0.3.8
discord.py==2.4.0
filelock==3.14.0
flake8==7.1.0
frozenlist==1.4.1
idna==3.7
isort==5.13.2
mccabe==0.7.0
mock==5.1.0
multidict==6.0.5
mypy==1.11.0
mypy-extensions==1.0.0
peewee==3.17.6
pillow==10.4.0
platformdirs==4.2.1
pycodestyle==2.12.0
pyflakes==3.2.0
pylint==3.2.6
python-dotenv==1.0.1
requests==2.32.3
retrying==1.3.4
ruff==0.5.5
six==1.16.0
tabulate==0.9.0
tomlkit==0.12.4
typed_ast==1.5.5
types-mock==5.1.0.20240425
types-Pillow==10.2.0.20240520
types-requests==2.32.0.20240712
types-tabulate==0.9.0.20240106
typing_extensions==4.11.0
update-checker==0.18.0
uritools==4.0.2
urlextract==1.9.0
urllib3==2.2.1
validators==0.33.0
yarl==1.9.4
